refactor: split satellite and launch provider modules
